requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.10
django-crispy-forms==2.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
googlemaps==4.10.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
numpy==1.26.4
openai==1.12.0
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
psycopg2-binary==2.9.9
pydantic==2.6.2
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
